requirements.txt
langchain
langchain_community
langchain_groq
python-dotenv
streamlit
arxiv
wikipedia
duckduckgo-search